# Improve Hubble decoding performance for drop, debug, policy and tracesock events