--- a/pkg/monitor/datapath_drop.go
+++ b/pkg/monitor/datapath_drop.go
@@ -9,6 +9,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/cilium/cilium/pkg/byteorder"
 	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/monitor/api"
 )
@@ -60,6 +61,33 @@ var sourceFileNames = map[int]string{
 	104: "icmp6.h",
 	105: "nodeport.h",
 	//end
+}
+
+// DecodeDropNotify will decode 'data' into the provided DropNotify structure
+func DecodeDropNotify(data []byte, dn *DropNotify) error {
+	return dn.decodeDropNotify(data)
+}
+
+func (n *DropNotify) decodeDropNotify(data []byte) error {
+	if l := len(data); l < DropNotifyLen {
+		return fmt.Errorf("unexpected DropNotify data length, expected %d but got %d", DropNotifyLen, l)
+	}
+
+	n.Type = data[0]
+	n.SubType = data[1]
+	n.Source = byteorder.Native.Uint16(data[2:4])
+	n.Hash = byteorder.Native.Uint32(data[4:8])
+	n.OrigLen = byteorder.Native.Uint32(data[8:12])
+	n.CapLen = byteorder.Native.Uint32(data[12:16])
+	n.SrcLabel = identity.NumericIdentity(byteorder.Native.Uint32(data[16:20]))
+	n.DstLabel = identity.NumericIdentity(byteorder.Native.Uint32(data[20:24]))
+	n.DstID = byteorder.Native.Uint32(data[24:28])
+	n.Line = byteorder.Native.Uint16(data[28:30])
+	n.File = data[30]
+	n.ExtError = int8(data[31])
+	n.Ifindex = byteorder.Native.Uint32(data[32:36])
+
+	return nil
 }
 
 func decodeBPFSourceFileName(fileId int) string {