Please ensure your pull request adheres to the following guidelines:

- [x] For first time contributors, read [Submitting a pull request](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#submitting-a-pull-request)
- [x] All code is covered by unit and/or runtime tests where feasible.
- [ ] All commits contain a well written commit description including a title,
      description and a `Fixes: #XXX` line if the commit addresses a particular
      GitHub issue.
- [ ] If your commit description contains a `Fixes: <commit-id>` tag, then
      please add the commit author[s] as reviewer[s] to this issue.
- [x] All commits are signed off. See the section [Developer’s Certificate of Origin](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#dev-coo)
- [ ] Provide a title or release-note blurb suitable for the release notes.
- [ ] Are you a user of Cilium? Please add yourself to the [Users doc](https://github.com/cilium/cilium/blob/main/USERS.md)
- [ ] Thanks for contributing!

<!-- Description of change -->
I implemented the function by referencing #24162.
reduced CPU usage.

DecodeDropNotify Benchmark Result
```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/monitor
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
=== RUN   BenchmarkNewDecodeDropNotify
BenchmarkNewDecodeDropNotify
BenchmarkNewDecodeDropNotify-4   	45821463	        27.00 ns/op	       0 B/op	       0 allocs/op
=== RUN   BenchmarkOldDecodeDropNotify
BenchmarkOldDecodeDropNotify
BenchmarkOldDecodeDropNotify-4   	 3343834	       368.0 ns/op	     144 B/op	       3 allocs/op
PASS
ok  	github.com/cilium/cilium/pkg/monitor	2.925s
```

DecodePolicyVerdictNotify
```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/monitor
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
=== RUN   BenchmarkNewDecodePolicyVerdictNotify
BenchmarkNewDecodePolicyVerdictNotify
BenchmarkNewDecodePolicyVerdictNotify-4         42731541                26.72 ns/op            0 B/op          0 allocs/op
=== RUN   BenchmarkOldDecodePolicyVerdictNotify
BenchmarkOldDecodePolicyVerdictNotify
BenchmarkOldDecodePolicyVerdictNotify-4          2977009               405.1 ns/op           112 B/op          3 allocs/op
PASS
ok      github.com/cilium/cilium/pkg/monitor    2.863s
```

DecodeDebugCapture/DecodeDebugMsg
```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/monitor
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
=== RUN   BenchmarkNewDecodeDebugCapture
BenchmarkNewDecodeDebugCapture
BenchmarkNewDecodeDebugCapture-4        64555694                18.92 ns/op            0 B/op          0 allocs/op
=== RUN   BenchmarkOldDecodeDebugCapture
BenchmarkOldDecodeDebugCapture
BenchmarkOldDecodeDebugCapture-4         4329416               272.1 ns/op            96 B/op          3 allocs/op
=== RUN   BenchmarkNewDecodeDebugMsg
BenchmarkNewDecodeDebugMsg
BenchmarkNewDecodeDebugMsg-4            74855434                16.42 ns/op            0 B/op          0 allocs/op
=== RUN   BenchmarkOldDecodeDebugMsg
BenchmarkOldDecodeDebugMsg
BenchmarkOldDecodeDebugMsg-4             4504531               264.1 ns/op            96 B/op          3 allocs/op
PASS
ok      github.com/cilium/cilium/pkg/monitor    5.501s
```

DecodeTraceSockNotify
```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/monitor
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
=== RUN   BenchmarkNewDecodeTraceSockNotify
BenchmarkNewDecodeTraceSockNotify
BenchmarkNewDecodeTraceSockNotify-4     92002668                12.99 ns/op            0 B/op          0 allocs/op
=== RUN   BenchmarkOldDecodeTraceSockNotify
BenchmarkOldDecodeTraceSockNotify
BenchmarkOldDecodeTraceSockNotify-4      2426386               439.2 ns/op           144 B/op          3 allocs/op
PASS
ok      github.com/cilium/cilium/pkg/monitor    2.862s
```
Fixes: #25657